주문 변수 타입, 병렬 주문시 버그 발생 해결

+++ src/finance/programgarden_finance/ls/overseas_stock/accno/COSOQ00201/blocks.py
@@ -425,15 +425,15 @@ class COSOQ00201OutBlock4(BaseModel):
     )
     """해외증권잔고구분코드명"""
 
-    AstkBalQty: float = Field(
-        default=0.0,
+    AstkBalQty: int = Field(
+        default=0,
         title="해외증권잔고수량",
         description="해외증권잔고수량"
     )
     """해외증권잔고수량"""
 
-    AstkSellAbleQty: float = Field(
-        default=0.0,
+    AstkSellAbleQty: int = Field(
+        default=0,
         title="해외증권매도가능수량",
         description="해외증권매도가능수량"
     )
